Fix boundary stitching
Usually we can tell these two cases by looking at whether edge i+1 has
the vertex of edge i as its source.  That's mostly, but not always the
case. :/

Assume the boundary consists of two chains beteen V1 and V2, and one of
the chains is on the CH.  The other chain is represented by the elements
in removal_boundary_vertices.  If V1 and V2 are connected by an edge,
and that's the edge we picked to represent V1 in the list, this will
blow up.

Therefore, we need to actually walk along the boundary before we
overwrite things.  Then we can just check if the right of us is removed
territoriy.